+++ src/main/java/com/planti/domain/community/dto/response/PostSummaryDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class PostSummaryDto { // 게시글 목록용 Dto
-    private Integer postId;
+    private long postId;
     private String title;
     private String content; // 글 내용
     private Long userId;
